Split MCP connector tests
